updated code to use the /jobs/jobTemplate API
It is preferable to use Job Templates instead of specifying print & mail options explicitly. If P&O options are not valid they give a low level error "sku not found" which is hard to debug. With Job Templates P&M options are guaranteed to be valid. Also they can be changed in the job template and no changes are needed in the code.